requirements.txt
fastapi
uvicorn[standard]
telethon
apscheduler
sqlalchemy
pydantic-settings
pyyaml
